Moves helper functions to own file #1074
